removed private access to go-netconf and instead used local existence. Consolidated go-netconf into netconf for easier readability
