+++ junosterraform/templates/go.mod.j2
@@ -7,11 +7,8 @@ require (
 	github.com/chrismarget/lambda-tf-registry v0.0.1
 	github.com/goreleaser/goreleaser v1.26.2
 	github.com/hashicorp/terraform-plugin-framework v1.15.0
-	github.com/vinpatel24/go-netconf v0.1.3
 	golang.org/x/crypto v0.39.0
 )
-
-replace github.com/vinpatel24/go-netconf => ./go-netconf
 
 require (
 	cloud.google.com/go v0.112.1 // indirect
