+++ terraform_provider/config.go
@@ -16,9 +16,8 @@
 
 package main
 
-import(
-//	"terraform-provider-junos-vsrx/netconf"
-        "github.com/Juniper/junos-terraform/terraform_providers/netconf"
+import (
+	"terraform_providers/netconf"
 )
 
 // Config is the configuration structure used to instantiate the Netconf provider.
