+++ terraform_provider/netconf/go-netconf/drivers/ssh/lowlevel/ssh.go
@@ -16,8 +16,8 @@ import (
 	"strings"
 	"time"
 
-	session "github.com/vinpatel24/go-netconf/session"
-	transport "github.com/vinpatel24/go-netconf/transport"
+	session "github.com/Juniper/go-netconf/session"
+	transport "github.com/Juniper/go-netconf/transport"
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/crypto/ssh/agent"
 )
