+++ terraform_provider/netconf/helper.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"sync"
 
-	driver "github.com/vinpatel24/go-netconf/drivers/driver"
-	sshdriver "github.com/vinpatel24/go-netconf/drivers/ssh"
+	driver "terraform_provider/netconf/go-netconf/drivers/driver"
+	sshdriver "terraform_provider/netconf/go-netconf/drivers/ssh"
 
 	"golang.org/x/crypto/ssh"
 )
